+++ client/src/components/fileComponent.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
 
 export function FileComponent({ fileInfo, index }) {
-    const { name, size } = fileInfo;
-  return ( //add file view using mimetypes
+  const { name, size } = fileInfo;
+  return (
+    //add file view using mimetypes
     <div>
-      <p>({index}) {name} ({size} KB)</p>
+      <p>
+        ({index}) {name} ({size} KB)
+      </p>
     </div>
   );
 }
-