+++ client/src/components/footer.jsx
@@ -5,13 +5,16 @@ import React from 'react';
 
 export function Footer() {
   return (
-    <div className='fixed bottom-0 w-full'>
+    <div className="fixed bottom-0 w-full">
       <footer className="footer bg-neutral-50 border-neutral-200 flex items-center justify-center sm:flex-wrap py-3 px-4">
-        <p className='text-center font-semibold mr-4'>Copyright ©️ 2024 droppr.</p>
-        <p className='text-center mr-4'>All rights reserved. </p>
-        <p className='text-center'>Built by Micah, Simon, Johnny, Nakul for Mountain Madness 2024</p>
+        <p className="text-center font-semibold mr-4">
+          Copyright ©️ 2024 droppr.
+        </p>
+        <p className="text-center mr-4">All rights reserved. </p>
+        <p className="text-center">
+          Built by Micah, Simon, Johnny, Nakul for Mountain Madness 2024
+        </p>
       </footer>
     </div>
-    
   );
 }