+++ client/src/components/header.jsx
@@ -11,7 +11,9 @@ export function Header() {
           <div className="flex items-center justify-between ml-12">
             <p
               className="text-3xl text-slate-700 font-bold hover:cursor-pointer"
-              onClick={() => { location.href = '/' }}
+              onClick={() => {
+                location.href = '/';
+              }}
             >
               droppr.
             </p>
@@ -19,13 +21,17 @@ export function Header() {
           <div className="flex items-center justify-between font-medium">
             <button
               className="mr-8 font-medium rounded-md bg-neutral-100 hover:bg-neutral-200 hover:cursor-pointer px-4 py-3 transition ease-in-out hover:scale-110"
-              onClick={() => { location.href = '/?view=about' }}
+              onClick={() => {
+                location.href = '/?view=about';
+              }}
             >
               About
             </button>
             <button
               className="mr-12 font-medium rounded-md bg-neutral-100 hover:bg-neutral-200 hover:cursor-pointer px-4 py-3 transition ease-in-out hover:scale-110"
-              onClick={() => { location.href = '/?view=drop' }}
+              onClick={() => {
+                location.href = '/?view=drop';
+              }}
             >
               Drop
             </button>