+++ client/src/components/index.js
@@ -2,5 +2,5 @@
 //
 
 export * from './header.jsx';
-export * from './footer.jsx'
-export * from './fileComponent.jsx'
+export * from './footer.jsx';
+export * from './fileComponent.jsx';