+++ client/workers/builder.js
@@ -1,0 +1,11 @@
+// Copyright (C) 2024 droppr. All rights reserved.
+//
+// workers/
+// builder.js
+
+// TODO: Store large files on the user's file system with IndexedDB, while
+//       receiving chunks; do this within this Web Worker.
+
+onmessage = (event) => {
+  // pass
+};